@@ -1,50 +1,40 @@
 {
   "name": "react-boilerplate",
   "private": true,
-  "version": "0.1.0",
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "tsc && vite build",
+    "build": "tsc -b && vite build",
+    "lint": "eslint .",
+    "format": "prettier --write \"**/*.{js,ts,tsx,md}\"",
     "prepare": "husky install",
-    "preview": "vite preview",
-    "lint": "eslint src --max-warnings=0",
-    "types": "tsc --project tsconfig.json --noEmit",
-    "format": "prettier --write \"**/*.{js,ts,tsx,md}\""
+    "preview": "vite preview"
   },
   "dependencies": {
-    "@hookform/resolvers": "^3.9.0",
-    "class-variance-authority": "^0.7.0",
-    "clsx": "^2.1.1",
-    "framer-motion": "^11.3.21",
+    "@radix-ui/react-dialog": "^1.1.2",
+    "class-variance-authority": "^0.7.1",
+    "classnames": "^2.5.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-hook-form": "^7.52.2",
-    "react-router-dom": "^6.25.1",
-    "tailwind-merge": "^2.4.0",
-    "yup": "^1.4.0"
+    "sass": "^1.81.0",
+    "vite-tsconfig-paths": "^5.1.3"
   },
   "devDependencies": {
-    "@faker-js/faker": "^8.4.1",
-    "@types/node": "^20.14.10",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "@typescript-eslint/eslint-plugin": "^7.16.1",
-    "@typescript-eslint/parser": "^7.16.1",
-    "@vitejs/plugin-react-swc": "^3.7.0",
-    "autoprefixer": "^10.4.19",
-    "eslint": "^8.56.0",
-    "eslint-plugin-prettier": "^5.1.3",
-    "eslint-plugin-react-hooks": "^4.6.2",
-    "eslint-plugin-react-refresh": "^0.4.7",
-    "eslint-plugin-tailwindcss": "^3.17.4",
-    "faker": "^6.6.6",
+    "@eslint/js": "^9.15.0",
+    "@types/node": "^22.10.0",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "@vitejs/plugin-react": "^4.3.4",
+    "eslint": "^9.15.0",
+    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-react-hooks": "^5.0.0",
+    "eslint-plugin-react-refresh": "^0.4.14",
+    "globals": "^15.12.0",
+    "prettier": "^3.4.1",
+    "typescript": "~5.6.2",
+    "typescript-eslint": "^8.15.0",
     "husky": "^9.0.11",
-    "postcss": "^8.4.39",
-    "prettier": "^3.3.3",
-    "prettier-plugin-tailwindcss": "0.6.5",
-    "tailwindcss": "^3.4.5",
-    "typescript": "^5.5.3",
-    "vite": "^5.3.3",
-    "vite-tsconfig-paths": "^4.3.2"
+    "vite": "^6.0.1"
   }
 }
