@@ -1,13 +1,17 @@
-import path from "path"
-import { defineConfig } from "vite"
-import react from "@vitejs/plugin-react-swc"
-import tsconfigPaths from "vite-tsconfig-paths"
+import path from 'path'
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react'
+import tsconfigPaths from 'vite-tsconfig-paths'
 
+const pathSrc = path.resolve(__dirname, "./src");
+
+// https://vite.dev/config/
 export default defineConfig({
 	plugins: [react(), tsconfigPaths()],
 	resolve: {
 		alias: {
-			"~": path.resolve(__dirname, "./src"),
+			'~': pathSrc,
 		},
 	},
+	base: '/repository-name/'
 })
